Make conversion from size_t to double explicit.
ConnConstRef is a tuple whose first value is double. Previously, candidateCrossingCount (which is a size_t) was being passed into std::make_pair directly. MSVC warns about this conversion, even though we know it should "just work", so using static_cast asserts our confidence and hushes the compiler warnings.

+++ cola/libavoid/router.cpp
@@ -1278,7 +1278,7 @@ class CrossingConnectorsInfo
             connSet.clear();
 
             // Return the candidate connector and its original crossing count.
-            return std::make_pair(candidateCrossingCount, candidateConnector);
+            return std::make_pair(static_cast<double>(candidateCrossingCount), candidateConnector);
         }
 
         // Returns the iterator to the group that the given conn is in,
